Removed duplicate email linking code - better error handling
