+++ app/src/hooks/usePrivyLinkEmail.ts
@@ -1,0 +1,70 @@
+import { toast } from "sonner"
+import { useLinkAccount, usePrivy, useUpdateAccount } from "@privy-io/react-auth"
+import { useRef } from "react"
+import { syncPrivyUser } from "@/db/privy"
+import { useHandlePrivyErrors } from "./useHandlePrivyErrors"
+import { useUser } from "./useUser"
+
+export const usePrivyEmail = (userId: string) => {
+
+    const isLinking = useRef(false)
+    const onError = useHandlePrivyErrors()
+    const { invalidate: invalidateUser } = useUser({ id: userId, enabled: false })
+    const { user: privyUser, unlinkEmail } = usePrivy()
+
+    const { linkEmail } = useLinkAccount({
+        onSuccess: async ({ user: updatedPrivyUser, linkMethod }) => {
+            if (linkMethod === "email" && isLinking.current) {
+                toast.promise(syncPrivyUser(updatedPrivyUser)
+                    .then(() => invalidateUser())
+                    .then(() => isLinking.current = false),
+                    {
+                        loading: "Adding email...",
+                        success: "Email added successfully",
+                        error: "Failed to add email",
+                    })
+            }
+        },
+        onError
+    })
+
+    const { updateEmail } = useUpdateAccount({
+        onSuccess: async ({ user: updatedPrivyUser, updateMethod }) => {
+            if (updateMethod === "email") {
+                toast.promise(syncPrivyUser(updatedPrivyUser)
+                    .then(() => invalidateUser()), {
+                    loading: "Updating email...",
+                    success: "Email updated successfully",
+                    error: "Failed to update email",
+                })
+            }
+        },
+        onError
+    })
+
+    const handleUnlinkEmail = () => {
+        if (privyUser?.email) {
+            toast.promise(unlinkEmail(privyUser.email.address), {
+                loading: "Deleting email...",
+                success: (updatedPrivyUser) => {
+                    syncPrivyUser(updatedPrivyUser)
+                        .then(() => invalidateUser())
+                        .then(() => "User invalidates")
+                    return "Email deleted successfully"
+                },
+                error: "Failed to delete email",
+            })
+        }
+    }
+
+    const linkEmailWithState = () => {
+        isLinking.current = true
+        linkEmail()
+    }
+
+    return {
+        linkEmail: linkEmailWithState,
+        updateEmail,
+        unlinkEmail: handleUnlinkEmail
+    }
+}
